Subvoxel r2s workflow support

+++ scripts/r2s.py
@@ -142,8 +142,11 @@ def step1():
         mesh = Mesh(structured=False, mesh=meshtal)
 
     ves = list(mesh.iter_ve())
+    tags_keep = ("cell_number", "cell_fracs",
+                 "cell_largest_frac_number", "cell_largest_frac")
     for tag in mesh.mesh.getAllTags(ves[0]):
-        mesh.mesh.destroyTag(tag, True)
+        if tag.name not in tags_keep:
+            mesh.mesh.destroyTag(tag, True)
     mesh.mesh.save('blank_mesh.h5m')
     print('The file blank_mesh.h5m has been saved to disk.')
     print('Do not delete this file; it is needed by r2s.py step2.\n')
@@ -161,6 +164,10 @@ def step2():
     tot_phtn_src_intensities = config.get('step2', 'tot_phtn_src_intensities')
     tag_name = "source_density"
 
+    if sub_voxel:
+        geom = config.get('step1', 'geom')
+        load(geom)
+        cell_mats = cell_materials(geom)
     h5_file = 'phtn_src.h5'
     if not isfile(h5_file):
         photon_source_to_hdf5('phtn_src')
@@ -169,7 +176,8 @@ def step2():
         print('Writing source for decay time: {0}'.format(dc))
         mesh = Mesh(structured=structured, mesh='blank_mesh.h5m')
         tags = {('TOTAL', dc): tag_name}
-        photon_source_hdf5_to_mesh(mesh, h5_file, tags)
+        photon_source_hdf5_to_mesh(mesh, h5_file, tags, sub_voxel=sub_voxel,
+                                   cell_mats=cell_mats)
         mesh.mesh.save('{0}_{1}.h5m'.format(output, i+1))
         intensity = total_photon_source_intensity(mesh, tag_name)
         intensities += "{0}: {1}\n".format(dc, intensity)
